fix : 도서, 도서별 저자 데이터 형 변경

+++ src/main/java/store/buzzbook/core/repository/product/BookAuthorRepository.java
@@ -4,9 +4,8 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import store.buzzbook.core.entity.product.Author;
 import store.buzzbook.core.entity.product.BookAuthor;
 
 public interface BookAuthorRepository extends JpaRepository<BookAuthor, Integer> {
-	List<BookAuthor> findAllByBookId(int bookId);
+	List<BookAuthor> findAllByBookId(long bookId);
 }
